@@ -9,11 +9,13 @@
 
 namespace BK2K\BootstrapPackage\Parser;
 
-use Leafo\ScssPhp\Compiler;
-use Leafo\ScssPhp\Formatter\Crunched;
-use Leafo\ScssPhp\Version;
+use ScssPhp\ScssPhp\Compiler;
+use ScssPhp\ScssPhp\Formatter\Crunched;
+use ScssPhp\ScssPhp\Version;
+use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Utility\PathUtility;
 
 /**
  * ScssParser
@@ -25,7 +27,7 @@ class ScssParser extends AbstractParser
      */
     public function __construct()
     {
-        if (!class_exists('Leafo\ScssPhp\Version', false)) {
+        if (!class_exists('ScssPhp\ScssPhp\Version')) {
             require_once ExtensionManagementUtility::extPath('bootstrap_package') . '/Contrib/scssphp/scss.inc.php';
         }
     }
@@ -83,12 +85,73 @@ class ScssParser extends AbstractParser
                 'sourceMapBasepath' => '<PATH DOES NOT EXIST BUT SUPRESSES WARNINGS>'
             ]);
         }
-        $css = $scss->compile('@import "' . $file . '"');
+        $absoluteFilename = $settings['file']['absolute'];
+        // Adds visual directory path of the initial file as import path
+        // This scenarios happens, when e.g. developing packages using the `path`
+        // repository feature of Composer - having one package in `public/typo3conf/ext/`
+        // and the other one symlinked in e.g. `packages/`.
+        // Since the PHP SCSS parser works on resolved real paths, the symlinked context is lost.
+        $visualImportPath = dirname($absoluteFilename);
+        $scss->addImportPath(function ($url) use ($visualImportPath) {
+            // Resolve potential back paths manually using PathUtility::getCanonicalPath,
+            // but make sure we do not break out of TYPO3 application path using GeneralUtility::getFileAbsFileName
+            // Also resolve EXT: paths if given
+            $isTypo3Absolute = strpos($url, 'EXT:') === 0;
+            $fileName = $isTypo3Absolute ? $url : $visualImportPath . '/' . $url;
+            $full = GeneralUtility::getFileAbsFileName(PathUtility::getCanonicalPath($fileName));
+            // The API forces us to check the existence of files paths, with or without url.
+            // We must only return a string if the file to be imported actually exists.
+            $hasExtension = preg_match('/[.]s?css$/', $url);
+            if (
+                is_file($file = $full . '.scss') ||
+                ($hasExtension && is_file($file = $full))
+            ) {
+                // We could trigger a deprecation message here at some point
+                return $file;
+            }
 
-        $absoluteFilename = GeneralUtility::getFileAbsFileName($file);
-        $relativePath = $settings['cache']['tempDirectoryRelativeToRoot'] . dirname(substr($absoluteFilename, strlen($this->getPathSite()))) . '/';
+            return null;
+        });
+        // Add extensions path to import paths, so that we can use paths relative to this directory to resolve imports
+        $scss->addImportPath(Environment::getExtensionsPath());
+
+        // Make paths in url() statements relative to site root
+        $absoluteFilePath = dirname($absoluteFilename);
+        $relativeFilePath = PathUtility::getAbsoluteWebPath($absoluteFilePath);
+        $absoluteBootstrapPackageThemePath =
+            ExtensionManagementUtility::extPath('bootstrap_package') . 'Resources/Public/Scss/Theme';
+        $relativeBootstrapPackageThemePath = PathUtility::getAbsoluteWebPath($absoluteBootstrapPackageThemePath);
+        $scss->registerFunction(
+            'url',
+            function ($args) use (
+                $scss,
+                $absoluteFilePath,
+                $relativeFilePath,
+                $absoluteBootstrapPackageThemePath,
+                $relativeBootstrapPackageThemePath
+            ) {
+                $marker = $args[0][1];
+                $args[0][1] = '';
+                $result = $scss->compileValue($args[0]);
+                if (substr_compare($result, 'data:', 0, 5, true) !== 0) {
+                    if (is_file(PathUtility::getCanonicalPath($absoluteFilePath . '/' . $result))) {
+                        $result = PathUtility::getCanonicalPath($relativeFilePath . '/' . $result);
+                    } elseif (is_file(PathUtility::getCanonicalPath($absoluteBootstrapPackageThemePath . '/' . $result))) {
+                        $result = PathUtility::getCanonicalPath($relativeBootstrapPackageThemePath . '/' . $result);
+                    }
+                    $result = substr($result, 0, 1) === '/' ? substr($result, 1) : $result;
+                }
+                return 'url(' . $marker . $result . $marker . ')';
+            }
+        );
+
+        // Compile file
+        $css = $scss->compile('@import "' . $absoluteFilename . '"');
+
+        // Fix paths in url() statements to be relative to temp directory
+        $relativeTempPath = $settings['cache']['tempDirectoryRelativeToRoot'];
         $search = '%url\s*\(\s*[\\\'"]?(?!(((?:https?:)?\/\/)|(?:data:?:)))([^\\\'")]+)[\\\'"]?\s*\)%';
-        $replace = 'url("' . $relativePath . '$3")';
+        $replace = 'url("' . $relativeTempPath . '$3")';
         $css = preg_replace($search, $replace, $css);
 
         return [
