@@ -66,10 +66,11 @@ class InstallService
                 'There is already an ' . $filename . ' configuration file in the root directory, '
                 . 'please make sure that configuration files are protected and the url rewritings are set properly. '
                 . 'An example configuration is located at: "typo3conf/ext/bootstrap_package/Configuration/Server/_' . $filename . '"',
-                'Webserver coniguration file "' . $filename . '" already exists',
+                'Webserver configuration file "' . $filename . '" already exists',
                 FlashMessage::NOTICE,
                 true
             );
+            /** @extensionScannerIgnoreLine */
             $this->addFlashMessage($flashMessage);
             return;
         }
@@ -81,10 +82,11 @@ class InstallService
             $flashMessage = GeneralUtility::makeInstance(
                 FlashMessage::class,
                 'For securing configuration files and optimization purposes an example ' . $filename . ' file was placed in your root directory.',
-                'Webserver coniguration file "' . $filename . '" was placed in the root directory.',
+                'Webserver configuration file "' . $filename . '" was placed in the root directory.',
                 FlashMessage::OK,
                 true
             );
+            /** @extensionScannerIgnoreLine */
             $this->addFlashMessage($flashMessage);
         }
     }
