@@ -1,4 +1,5 @@
 <?php
+declare(strict_types = 1);
 
 /*
  * This file is part of the package bk2k/bootstrap-package.
@@ -13,28 +14,54 @@ use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Install\Updates\DatabaseUpdatedPrerequisite;
+use TYPO3\CMS\Install\Updates\RepeatableInterface;
+use TYPO3\CMS\Install\Updates\UpgradeWizardInterface;
 
 /**
  * CarouselItemTypeUpdate
  */
-class CarouselItemTypeUpdate extends \TYPO3\CMS\Install\Updates\AbstractUpdate
+class CarouselItemTypeUpdate implements UpgradeWizardInterface, RepeatableInterface
 {
     /**
-     * @var string
+     * @return string
      */
-    protected $title = '[BootstrapPackage] Migrate carousel item types';
+    public function getIdentifier(): string
+    {
+        return self::class;
+    }
 
     /**
-     * Checks if an update is needed
-     *
-     * @param string &$description The description for the update
-     * @return bool Whether an update is needed (TRUE) or not (FALSE)
+     * @return string
      */
-    public function checkForUpdate(&$description)
+    public function getTitle(): string
     {
-        if ($this->isWizardDone()) {
-            return false;
-        }
+        return '[Bootstrap Package] Migrate carousel item types';
+    }
+
+    /**
+     * @return string
+     */
+    public function getDescription(): string
+    {
+        return '';
+    }
+
+    /**
+     * @return array
+     */
+    public function getPrerequisites(): array
+    {
+        return [
+            DatabaseUpdatedPrerequisite::class
+        ];
+    }
+
+    /**
+     * @return bool
+     */
+    public function updateNecessary(): bool
+    {
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('tx_bootstrappackage_carousel_item');
         $queryBuilder->getRestrictions()->removeAll()->add(GeneralUtility::makeInstance(DeletedRestriction::class));
         $elementCount = $queryBuilder->count('uid')
@@ -53,13 +80,9 @@ class CarouselItemTypeUpdate extends \TYPO3\CMS\Install\Updates\AbstractUpdate
     }
 
     /**
-     * Performs the database update
-     *
-     * @param array &$databaseQueries Queries done in this update
-     * @param string &$customMessage Custom message
      * @return bool
      */
-    public function performUpdate(array &$databaseQueries, &$customMessage)
+    public function executeUpdate(): bool
     {
         $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable('tx_bootstrappackage_carousel_item');
         $queryBuilder = $connection->createQueryBuilder();
@@ -86,16 +109,12 @@ class CarouselItemTypeUpdate extends \TYPO3\CMS\Install\Updates\AbstractUpdate
                     )
                 )
                 ->set('item_type', $this->mapValues($record['item_type']));
-            $databaseQueries[] = $queryBuilder->getSQL();
             $queryBuilder->execute();
         }
-        $this->markWizardAsDone();
         return true;
     }
 
     /**
-     * Map the old to the new values
-     *
      * @param int $type
      * @return string
      */
