@@ -14,17 +14,15 @@ use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Install\Updates\DatabaseUpdatedPrerequisite;
+use TYPO3\CMS\Install\Updates\RepeatableInterface;
+use TYPO3\CMS\Install\Updates\UpgradeWizardInterface;
 
 /**
  * FrameClassToBackgroundUpdate
  */
-class FrameClassToBackgroundUpdate extends \TYPO3\CMS\Install\Updates\AbstractUpdate
+class FrameClassToBackgroundUpdate implements UpgradeWizardInterface, RepeatableInterface
 {
-    /**
-     * @var string
-     */
-    protected $title = '[BootstrapPackage] Migrate obsolete frame_class options to background color';
-
     /**
      * @var string
      */
@@ -44,16 +42,44 @@ class FrameClassToBackgroundUpdate extends \TYPO3\CMS\Install\Updates\AbstractUp
     ];
 
     /**
-     * Checks if an update is needed
-     *
-     * @param string &$description The description for the update
-     * @return bool Whether an update is needed (TRUE) or not (FALSE)
+     * @return string
      */
-    public function checkForUpdate(&$description)
+    public function getIdentifier(): string
     {
-        if ($this->isWizardDone()) {
-            return false;
-        }
+        return self::class;
+    }
+
+    /**
+     * @return string
+     */
+    public function getTitle(): string
+    {
+        return '[Bootstrap Package] Migrate obsolete frame_class options to background color';
+    }
+
+    /**
+     * @return string
+     */
+    public function getDescription(): string
+    {
+        return '';
+    }
+
+    /**
+     * @return array
+     */
+    public function getPrerequisites(): array
+    {
+        return [
+            DatabaseUpdatedPrerequisite::class
+        ];
+    }
+
+    /**
+     * @return bool
+     */
+    public function updateNecessary(): bool
+    {
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($this->table);
         $queryBuilder->getRestrictions()->removeAll()->add(GeneralUtility::makeInstance(DeletedRestriction::class));
         $elementCount = $queryBuilder->count('uid')
@@ -72,13 +98,9 @@ class FrameClassToBackgroundUpdate extends \TYPO3\CMS\Install\Updates\AbstractUp
     }
 
     /**
-     * Performs the database update
-     *
-     * @param array &$databaseQueries Queries done in this update
-     * @param string &$customMessage Custom message
      * @return bool
      */
-    public function performUpdate(array &$databaseQueries, &$customMessage)
+    public function executeUpdate(): bool
     {
         $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable($this->table);
         $queryBuilder = $connection->createQueryBuilder();
@@ -112,16 +134,12 @@ class FrameClassToBackgroundUpdate extends \TYPO3\CMS\Install\Updates\AbstractUp
                     'background_color_class',
                     $this->mapValues($record[$this->field])
                 );
-            $databaseQueries[] = $queryBuilder->getSQL();
             $queryBuilder->execute();
         }
-        $this->markWizardAsDone();
         return true;
     }
 
     /**
-     * Map the old to the new values
-     *
      * @param string|int $value
      * @return string|null
      */
