@@ -10,8 +10,8 @@
 namespace BK2K\BootstrapPackage\ViewHelpers\Data;
 
 use BK2K\BootstrapPackage\Utility\ImageVariantsUtility;
-use TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper;
 use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
+use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithRenderStatic;
 
 /**
@@ -27,11 +27,12 @@ class ImageVariantsViewHelper extends AbstractViewHelper
     public function initializeArguments()
     {
         parent::initializeArguments();
-        $this->registerArgument('as', 'string', 'Name of variable to create', true);
-        $this->registerArgument('variants', 'array', 'Variants for responsive images', false);
-        $this->registerArgument('multiplier', 'array', 'Multiplier to calculate responsive image widths', false);
-        $this->registerArgument('gutters', 'array', 'Gutter that needs to be respected when calculating responsive image widths', false);
-        $this->registerArgument('corrections', 'array', 'Corrections to be applied after calculationof image widths', false);
+        $this->registerArgument('as', 'string', 'Name of variable to create.', true);
+        $this->registerArgument('variants', 'array', 'Variants for responsive images.', false);
+        $this->registerArgument('multiplier', 'array', 'Multiplier to calculate responsive image widths.', false);
+        $this->registerArgument('gutters', 'array', 'Gutter that needs to be respected when calculating responsive image widths.', false);
+        $this->registerArgument('corrections', 'array', 'Corrections to be applied after calculation of image widths.', false);
+        $this->registerArgument('aspectRatio', 'float', 'Set aspect ratio for all variants.', false);
     }
 
     /**
@@ -45,7 +46,7 @@ class ImageVariantsViewHelper extends AbstractViewHelper
         \Closure $renderChildrenClosure,
         RenderingContextInterface $renderingContext
     ) {
-        $variants = ImageVariantsUtility::getImageVariants($arguments['variants'], $arguments['multiplier'], $arguments['gutters'], $arguments['corrections']);
+        $variants = ImageVariantsUtility::getImageVariants($arguments['variants'], $arguments['multiplier'], $arguments['gutters'], $arguments['corrections'], $arguments['aspectRatio']);
         $renderingContext->getVariableProvider()->add($arguments['as'], $variants);
     }
 }
