@@ -43,7 +43,10 @@ return [
         'showRecordFieldList' => '
             hidden,
             tt_content,
-            header
+            header,
+            subheader,
+            bodytext,
+            icon_file,
         ',
     ],
     'types' => [
@@ -52,6 +55,8 @@ return [
                 --palette--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:palette.general;general,
                 header,
                 subheader,
+                bodytext,
+                link,
                 icon_file,
                 --div--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:tabs.access,
                 --palette--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:palette.visibility;visibility,
@@ -94,7 +99,7 @@ return [
                 'type' => 'select',
                 'renderType' => 'selectSingle',
                 'foreign_table' => 'tt_content',
-                'foreign_table_where' => 'AND tt_content.pid=###CURRENT_PID### AND tt_content.CType="icon_group"',
+                'foreign_table_where' => 'AND tt_content.pid=###CURRENT_PID### AND tt_content.{#CType}=\'icon_group\'',
                 'maxitems' => 1,
                 'default' => 0,
             ],
@@ -183,7 +188,6 @@ return [
             ]
         ],
         'header' => [
-            'exclude' => true,
             'label' => 'LLL:EXT:bootstrap_package/Resources/Private/Language/Backend.xlf:icon_group_item.header',
             'config' => [
                 'type' => 'input',
@@ -192,7 +196,6 @@ return [
             ],
         ],
         'subheader' => [
-            'exclude' => true,
             'label' => 'LLL:EXT:bootstrap_package/Resources/Private/Language/Backend.xlf:icon_group_item.subheader',
             'config' => [
                 'type' => 'input',
@@ -200,8 +203,38 @@ return [
                 'eval' => 'trim'
             ],
         ],
+        'bodytext' => [
+            'label' => 'LLL:EXT:bootstrap_package/Resources/Private/Language/Backend.xlf:icon_group_item.bodytext',
+            'l10n_mode' => 'prefixLangTitle',
+            'l10n_cat' => 'text',
+            'config' => [
+                'type' => 'text',
+                'cols' => '80',
+                'rows' => '15',
+                'softref' => 'typolink_tag,images,email[subst],url',
+                'enableRichtext' => true
+            ],
+        ],
+        'link' => [
+            'label' => 'LLL:EXT:bootstrap_package/Resources/Private/Language/Backend.xlf:icon_group_item.link',
+            'config' => [
+                'type' => 'input',
+                'renderType' => 'inputLink',
+                'size' => 50,
+                'max' => 1024,
+                'eval' => 'trim',
+                'fieldControl' => [
+                    'linkPopup' => [
+                        'options' => [
+                            'title' => 'LLL:EXT:bootstrap_package/Resources/Private/Language/Backend.xlf:icon_group_item.link',
+                        ],
+                    ],
+                ],
+                'softref' => 'typolink'
+            ],
+            'l10n_mode' => 'exclude',
+        ],
         'icon_file' => [
-            'exclude' => true,
             'label' => 'LLL:EXT:bootstrap_package/Resources/Private/Language/Backend.xlf:icon_group_item.icon_file',
             'config' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::getFileFieldTCAConfig(
                 'icon_file',
