@@ -104,7 +104,7 @@ return [
                 'type' => 'select',
                 'renderType' => 'selectSingle',
                 'foreign_table' => 'tt_content',
-                'foreign_table_where' => 'AND tt_content.pid=###CURRENT_PID### AND tt_content.CType="tab"',
+                'foreign_table_where' => 'AND tt_content.pid=###CURRENT_PID### AND tt_content.{#CType}=\'tab\'',
                 'maxitems' => 1,
                 'default' => 0,
             ],
@@ -210,8 +210,7 @@ return [
                 'cols' => '80',
                 'rows' => '15',
                 'softref' => 'typolink_tag,images,email[subst],url',
-                'enableRichtext' => true,
-                'richtextConfiguration' => 'default'
+                'enableRichtext' => true
             ],
         ],
         'media' => [
