@@ -2,14 +2,14 @@
 /**
  * SCSSPHP
  *
- * @copyright 2017 Leaf Corcoran
+ * @copyright 2015-2019 Leaf Corcoran
  *
  * @license http://opensource.org/licenses/MIT MIT
  *
- * @link http://leafo.github.io/scssphp
+ * @link http://scssphp.github.io/scssphp
  */
 
-namespace Leafo\ScssPhp\Base;
+namespace ScssPhp\ScssPhp\Base;
 
 /**
  * Range
